+++ packages/turbo/src/lib/index.ts
@@ -1,0 +1,3 @@
+import * as logger from "./logger.js";
+
+export { logger };