+++ packages/$private/configs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @serwist/configs
 
+## 9.1.0
+
+### Patch Changes
+
+- [`852df26`](https://github.com/serwist/serwist/commit/852df2609f700d28de6433e0cb6669ade13c5b95) Thanks [@DuCanhGH](https://github.com/DuCanhGH)! - chore(deps): monthly dependencies maintenance
+
+## 9.0.15
+
+### Patch Changes
+
+- [`de27be5`](https://github.com/serwist/serwist/commit/de27be5c8c48afc6122e046dc116696cedfc93e4) Thanks [@DuCanhGH](https://github.com/DuCanhGH)! - chore(deps): monthly dependencies maintenance
+
+  - Just the regular stuff. Serwist 10 is still on the way!
+
 ## 10.0.0-preview.8
 
 ## 10.0.0-preview.7